Animation is workink and fight is wokring
Spriter is fully integrated into Engine, we can now make animations with
it and import it in the game.
Early concept art of background is finished. day and night in the desert
plus moving clouds.
Fight is working, the assassin can dash forward to attack certain
enemies in different locations, attack them and return to his original
location